requirements.txt
# 🌐 FastAPI et serveur ASGI (Uvicorn)
fastapi
uvicorn

# 🗄️ ORM SQLAlchemy Asynchrone
sqlalchemy
aiomysql
asyncpg
alembic
pymysql

# 🔐 Sécurité et cryptographie
passlib
passlib[scrypt]  # Utilisation de Scrypt au lieu de Bcrypt
cryptography

# 🔑 Authentification JWT
pyjwt

# 🔧 Gestion des variables d’environnement
python-dotenv

# ⚙️ Gestion des événements asynchrones
greenlet